app/requirements.txt
anyio==3.6.2
autopep8==2.0.1
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
ecdsa==0.18.0
fastapi==0.92.0
h11==0.14.0
httptools==0.5.0
idna==3.4
install==1.3.5
passlib==1.7.4
pyasn1==0.4.8
pycodestyle==2.10.0
pydantic==1.10.4
pymongo==4.3.3
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.4.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
